% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc}
\alias{calc}
\title{Title}
\usage{
calc(x, hill_params, ...)
}
\arguments{
\item{x}{a GeoTox object}

\item{hill_params}{output from \link{fit_hill}}

\item{...}{additional arguments passed to other functions}
}
\value{
The same object with additional fields added
}
\description{
Compute
}
\examples{
x <- simulate(GeoTox(),
              age = split(geo_tox_data$age, ~FIPS)[1:5],
              obesity = geo_tox_data$obesity[1:5, ],
              exposure = split(geo_tox_data$exposure, ~FIPS)[1:5],
              simulated_css = geo_tox_data$simulated_css,
              n = 10)
hill_params <- fit_hill(split(geo_tox_data$dose_response, ~casn))
x <- calc(x, hill_params)
}
\seealso{
\link{fit_hill}, \link{calc_internal_dose}, \link{calc_invitro_concentration},
\link{calc_concentration_response}
}
