% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dose_response.R
\name{calc_dose_response}
\alias{calc_dose_response}
\title{Title}
\usage{
calc_dose_response(resp, dose)
}
\arguments{
\item{resp}{data frame with columns "tp", "logAC50", "resp_max", "logc_min",
"logc_max".}

\item{dose}{dose}
}
\value{
data frame
}
\description{
Title
}
