% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_independent_action.R
\name{calc_independent_action}
\alias{calc_independent_action}
\title{Independent action prediction from the tcpl Hill function}
\usage{
calc_independent_action(Ci, tp, AC50, Emax, slope = 1)
}
\arguments{
\item{Ci}{individual chemical concentrations in regular space}

\item{tp}{top asymptotes}

\item{AC50}{AC50}

\item{Emax}{maximum mixture response}

\item{slope}{hill coefficient}
}
\value{
response value
}
\description{
Independent action prediction from the tcpl Hill function
}
