% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_response.R
\name{calc_response}
\alias{calc_response}
\title{Calculate concentrations and risks}
\usage{
calc_response(GT, overwrite = FALSE, ...)
}
\arguments{
\item{GT}{GeoTox object.}

\item{overwrite}{Logical indicating whether to overwrite existing values
(default FALSE).}

\item{...}{Additional arguments passed to \code{\link[=calc_risk]{calc_risk()}}.}
}
\value{
The same GeoTox object, invisibly.
}
\description{
Calculate internal dose, in vitro concentration, and risk estimates.
}
\details{
This is a wrapper around several other functions:
\itemize{
\item \code{\link[=calc_internal_dose]{calc_internal_dose()}} to calculate internal dose (D_int).
\item \code{\link[=calc_invitro_concentration]{calc_invitro_concentration()}} to calculate in vitro concentration
(C_invitro).
\item \code{\link[=calc_risk]{calc_risk()}} to calculate risk estimates.
}

If a \code{risk_name} argument is provided to \code{...} and it is not "risk", then
sensitivity analysis is assumed and the \code{sensitivity} argument in the
internal dose and in vitro concentration calculations is set to TRUE.
}
\examples{
\dontshow{
  tempdir <- withr::local_tempdir()
  op <- options(duckdb.home = tempdir, duckdb.extension_directory = tempdir)
}
# Setup required tables
sample_df <- tibble::tribble(
  ~FIPS, ~age, ~weight,
  10000, 25, "Normal",
  10000, 35,  "Obese",
  20000, 50, "Normal"
)
exposure_df <- tibble::tribble(
  ~FIPS, ~casn, ~route, ~mean, ~sd,
  10000, "00-00-1", "inhalation", 10, 1,
  10000, "00-00-2", "inhalation", 20, 1,
  20000, "00-00-1", "inhalation", 30, 1,
  20000, "00-00-2", "inhalation", 40, 1
)
css_df <- tibble::tribble(
  ~casn, ~age_lb, ~age_ub, ~weight, ~css,
  "00-00-1",  0, 49, "Normal", 21,
  "00-00-1", 50, 99, "Normal", 22,
  "00-00-1",  0, 49,  "Obese", 61,
  "00-00-1", 50, 99,  "Obese", 62,
  "00-00-2",  0, 49, "Normal", 11,
  "00-00-2", 50, 99, "Normal", 12,
  "00-00-2",  0, 49,  "Obese", 31,
  "00-00-2", 50, 99,  "Obese", 32
)
hill_df <- tibble::tribble(
  ~assay, ~casn, ~logc, ~resp,
  "a1", "00-00-1",    0,  10,
  "a1", "00-00-1",    1,  20,
  "a1", "00-00-1",    2,  80,
  "a1", "00-00-1",    3, 100,
  "a1", "00-00-2", -0.5,   5,
  "a1", "00-00-2",  0.5,  20,
  "a1", "00-00-2",  1.5,  55,
  "a1", "00-00-2",  2.5,  60
)
GT <- GeoTox() |>
  set_sample(sample_df) |>
  set_simulated_css(css_df) |>
  add_exposure_rate_params() |>
  add_hill_params(fit_hill(hill_df, assay = "assay", substance = "casn")) |>
  simulate_population(exposure = exposure_df)

# Calculate concentrations and risk
GT <- GT |> calc_response()

# Open a connection to GeoTox database
con <- get_con(GT)

# Look at relevant table

dplyr::tbl(con, "concentration") |> dplyr::collect()

dplyr::tbl(con, "risk") |> dplyr::collect()

# Clean up example
DBI::dbDisconnect(con)
file.remove(GT$db_info$dbdir)
\dontshow{
  options(op)
}
}
\seealso{
\code{\link[=calc_internal_dose]{calc_internal_dose()}}, \code{\link[=calc_invitro_concentration]{calc_invitro_concentration()}},
\code{\link[=calc_risk]{calc_risk()}}
}
