% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lengths.R
\name{.check_lengths}
\alias{.check_lengths}
\title{Check for matching lengths}
\usage{
.check_lengths(x, y)
}
\arguments{
\item{x}{list of arrays}

\item{y}{list of arrays}
}
\value{
boolean, TRUE for error
}
\description{
Check for matching lengths
}
\keyword{internal}
