% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{.check_names}
\alias{.check_names}
\title{Check for names}
\usage{
.check_names(x, names)
}
\arguments{
\item{x}{object to check}

\item{names}{names to look for}
}
\value{
boolean, TRUE for error
}
\description{
Check for names
}
