% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_types.R
\name{.check_types}
\alias{.check_types}
\title{Check types}
\usage{
.check_types(x, types, msg = "Incorrect type")
}
\arguments{
\item{x}{object to check}

\item{types}{allowed types}

\item{msg}{error message}
}
\value{
list version of input
}
\description{
Check types
}
\keyword{internal}
