% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hill_params.R
\name{extract_hill_params}
\alias{extract_hill_params}
\title{Extract results from 2-parameter fit_hill outputs}
\usage{
extract_hill_params(fits)
}
\arguments{
\item{fits}{list of outputs from fit_hill(..., fixed_slope = TRUE)}
}
\value{
data frame
}
\description{
Extract results from 2-parameter fit_hill outputs
}
\examples{
df <- data.frame(
  logc = -3:3,
  resp = 5 / (1 + 10^(1.2 * (0.4 - rep(-3:3, each = 3)))) + rnorm(21)
 )
fit <- fit_hill(df$logc, df$resp)
extract_hill_params(list(fit))

df_list <- list(df, df)
fits <- lapply(df_list, function(df) {fit_hill(df$logc, df$resp)})
extract_hill_params(fits)
}
