% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hill_2par.R
\name{fit_hill_2par}
\alias{fit_hill_2par}
\title{Fit 2-parameter Hill model}
\usage{
fit_hill_2par(log10_conc, resp)
}
\arguments{
\item{log10_conc}{base-10 log scale concentration}

\item{resp}{response}
}
\value{
fit and other stats
}
\description{
Fit 2-parameter Hill model
}
