% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_tox_data}
\alias{geo_tox_data}
\title{Geo Tox data}
\format{
A list with 4 items:
\describe{
\item{ice}{Description}
\item{MC_iter}{Description}
\item{cyp1a1}{Description}
\item{age}{Description}
\item{Css}{Description}
}
Css is a list with 5 items:
\describe{
\item{Css$age}{Description}
\item{Css$obesity}{Description}
\item{Css$httk}{Description}
\item{Css$'dose-resp'}{Description}
\item{Css$'ext-conc'}{Description}
}
}
\usage{
geo_tox_data
}
\description{
A subset of data from the GeoToxMIE source scripts
}
\keyword{datasets}
