% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTox-package.R
\docType{package}
\name{GeoTox-package}
\alias{GeoTox-package}
\title{GeoTox: Spatiotemporal mixture risk assessment}
\description{
Open-source software for connecting spatiotemporal exposure to individual and population-level risk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://niehs.github.io/GeoTox/ https://github.com/NIEHS/GeoTox}
}

}
\author{
\strong{Maintainer}: Kyle Messier \email{kyle.messier@nih.gov} [contributor]

Authors:
\itemize{
  \item Skylar Marvel \email{skylar.marvel@nih.gov} [contributor]
  \item David Reif \email{david.reif@nih.gov} [contributor]
}

}
\keyword{internal}
