% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixed_obesity.R
\name{get_fixed_obesity}
\alias{get_fixed_obesity}
\title{Title}
\usage{
get_fixed_obesity(obesity, simulated_css)
}
\arguments{
\item{obesity}{x}

\item{simulated_css}{x}
}
\value{
x
}
\description{
Title
}
