% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_ECx.R
\name{obj_ECx}
\alias{obj_ECx}
\title{A Generalized concentration addition style solution}
\usage{
obj_ECx(ECmix, E, Ci, tp, AC50)
}
\arguments{
\item{ECmix}{effective concentration of the mixture}

\item{E}{individual chemical responses}

\item{Ci}{individual chemical concentrations in regular space}

\item{tp}{top asymptotes}

\item{AC50}{AC50s}
}
\value{
objective value
}
\description{
Find the effective concentration of a mixture via an objective function
given the concentrations and inverse, based on a regular space AC50 and
concentrations.
}
