% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_GCA.R
\name{obj_GCA}
\alias{obj_GCA}
\title{Generalized concentration addition objective function}
\usage{
obj_GCA(effic, Ci, tp, AC50)
}
\arguments{
\item{effic}{natural log of individual chemical responses}

\item{Ci}{individual chemical concentrations in regular space}

\item{tp}{top asymptote}

\item{AC50}{AC50}
}
\value{
objective value
}
\description{
Use to find the optimal efficacy value, E, based on a regular space AC50 and
concentrations.
}
