% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_hill.R
\name{obj_hill}
\alias{obj_hill}
\title{Objective function for 2- or 3-parameter Hill model}
\usage{
obj_hill(par, log10_conc, resp)
}
\arguments{
\item{par}{parameters}

\item{log10_conc}{base-10 log scale concentration}

\item{resp}{response}
}
\value{
value
}
\description{
Objective function for 2- or 3-parameter Hill model
}
