% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_hill_2par.R
\name{obj_hill_2par}
\alias{obj_hill_2par}
\title{Objective function for 2-parameter Hill model}
\usage{
obj_hill_2par(p, log10_conc, resp)
}
\arguments{
\item{p}{parameters c("tp", "log10_ga", "err")}

\item{log10_conc}{base-10 log scale concentration}

\item{resp}{response}
}
\value{
value
}
\description{
Objective function for 2-parameter Hill model
}
