% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_Css.R
\name{sample_Css}
\alias{sample_Css}
\title{Sample from pre-generated C_ss data}
\usage{
sample_Css(simulated_css, age, obesity)
}
\arguments{
\item{simulated_css}{list of pre-generated C_ss data, for details see:
\code{vignette("package_data", package = "GeoTox")}.}

\item{age}{list or atomic vector of ages.}

\item{obesity}{list or atomic vector of obesity status.}
}
\value{
list of matrices containing C_ss values. Columns are sorted to have
consistent order across functions.
}
\description{
Sample from pre-generated C_ss data
}
\examples{
# Vector inputs
sample_Css(geo_tox_data$simulated_css,
           c(15, 25, 35),
           c("Normal", "Obese", "Normal"))

# List inputs
sample_Css(geo_tox_data$simulated_css,
           list(c(34, 29), 55),
           list(c("Obese", "Normal"), "Normal"))
}
