% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boundaries.R
\name{set_boundaries}
\alias{set_boundaries}
\title{Title}
\usage{
set_boundaries(x, region = NULL, group = NULL)
}
\arguments{
\item{x}{.}

\item{region}{.}

\item{group}{.}
}
\value{
.
}
\description{
Title
}
