% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplHillConc.R
\name{tcplHillConc}
\alias{tcplHillConc}
\title{TCPL Hill model}
\usage{
tcplHillConc(resp, tp, ga, gw)
}
\arguments{
\item{resp}{response value}

\item{tp}{top asymptote}

\item{ga}{AC50}

\item{gw}{hill coefficient}
}
\value{
concentration in regular space
}
\description{
Calculate the concentration in regular space for a given value
}
