% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcplHillVal.R
\name{tcplHillVal}
\alias{tcplHillVal}
\title{TCPL Hill model}
\usage{
tcplHillVal(conc, tp, ga, gw, bt = 0)
}
\arguments{
\item{conc}{concentration in regular space}

\item{tp}{top asymptote}

\item{ga}{AC50}

\item{gw}{hill coefficient}

\item{bt}{bottom asymptote}
}
\value{
response value
}
\description{
Calculate the efficacy for a given concentration in regular space
}
