% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_concentration_response.R
\name{calc_concentration_response}
\alias{calc_concentration_response}
\title{Title}
\usage{
calc_concentration_response(resp, concentration)
}
\arguments{
\item{resp}{data frame with columns "tp", "logAC50", "resp_max", "logc_min",
"logc_max".}

\item{concentration}{concentration}
}
\value{
data frame
}
\description{
Title
}
