% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_concentration_response.R
\name{calc_concentration_response}
\alias{calc_concentration_response}
\title{Calculate the mixture response from one of three different approaches: IA, GCA ,or Hazard Quotient}
\usage{
calc_concentration_response(resp, concentration, fixed = FALSE)
}
\arguments{
\item{resp}{data frame with columns "tp", "logAC50", "resp_max", "logc_min",
"logc_max".}

\item{concentration}{concentration}

\item{fixed}{}
}
\value{
data frame
}
\description{
Calculate the combined response of multiple chemicals. It calculates the generalized concentration addition response, the independent actionr response, and a hazard quotient
}
