% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_invitro_concentration.R
\name{calc_invitro_concentration}
\alias{calc_invitro_concentration}
\title{Calculate \emph{in vitro} concentration}
\usage{
calc_invitro_concentration(D_int, C_ss = NULL)
}
\arguments{
\item{D_int}{internal chemical dose in \eqn{\frac{mg}{kg}}}

\item{C_ss}{steady-state plasma concentration in \eqn{\frac{\mu M}{mg / kg}}}
}
\value{
\emph{in vitro} equivalent plasma concentration in \eqn{\mu M}
}
\description{
TODO A short description...
}
\details{
TODO Additional details...
\deqn{C_{plasma} = C_{ss} \,\times\, D_{int}}

TODO If C_ss is not provided, uses \emph{httk} to create...
}
