% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_response.R
\name{calculate_response}
\alias{calculate_response}
\title{Calculate response}
\usage{
calculate_response(x, ...)
}
\arguments{
\item{x}{GeoTox object}

\item{...}{additional arguments passed to other functions. See details.}
}
\value{
The same object with additional fields added or updated
}
\description{
Calculate mixture response for GeoTox population data
}
\details{
Additional parameters include \code{time}, \code{BW}, and \code{scaling} for
\link{calc_internal_dose}, and \code{max_mult} for \link{calc_concentration_response}.
}
\seealso{
\link{calc_internal_dose}, \link{calc_invitro_concentration},
\link{calc_concentration_response}
}
