% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_response.R
\name{calculate_response}
\alias{calculate_response}
\title{Calculate response}
\usage{
calculate_response(x, ...)
}
\arguments{
\item{x}{GeoTox object}

\item{...}{additional arguments passed to other functions. See details.}
}
\value{
The same object with additional fields added or updated
}
\description{
Calculate mixture response for GeoTox population data
}
\details{
Additional parameters include \code{time}, \code{BW}, and \code{scaling} for
\link{calc_internal_dose}, and \code{max_mult} for \link{calc_concentration_response}.
}
\examples{
# Create GeoTox object
geoTox <- GeoTox()

# The following fields are required inputs
geoTox$IR <- 2
geoTox$C_ext <- matrix(3)
geoTox$C_ss <- 5
geoTox$hill_params <- fit_hill(data.frame(logc = c(-1, 0, 1),
                                          resp = c(10, 5, 0)))

# Calculate response
geoTox <- calculate_response(geoTox)

# The following fields will be computed
geoTox$D_int
geoTox$C_invitro
geoTox$resp
}
\seealso{
\link{calc_internal_dose}, \link{calc_invitro_concentration},
\link{calc_concentration_response}
}
