% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_response.R
\name{calculate_response}
\alias{calculate_response}
\title{Calculate response}
\usage{
calculate_response(x, hill_params, ...)
}
\arguments{
\item{x}{GeoTox object}

\item{hill_params}{output from \link{fit_hill}}

\item{...}{additional arguments passed to other functions}
}
\value{
The same object with additional fields added or updated
}
\description{
Calculate mixture response for GeoTox population data
}
\examples{
x <- simulate_population(
  GeoTox(),
  age = split(geo_tox_data$age, ~FIPS)[1:5],
  obesity = geo_tox_data$obesity[1:5, ],
  exposure = split(geo_tox_data$exposure, ~FIPS)[1:5],
  simulated_css = geo_tox_data$simulated_css,
  n = 10
)
hill_params <- fit_hill(split(geo_tox_data$dose_response, ~casn))
x <- calculate_response(x, hill_params)
}
\seealso{
\link{fit_hill}, \link{calc_internal_dose}, \link{calc_invitro_concentration},
\link{calc_concentration_response}
}
