% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sensitivity.R
\name{compute_sensitivity}
\alias{compute_sensitivity}
\title{Title}
\usage{
compute_sensitivity(
  x,
  vary = c("age", "obesity", "css_params", "fit_params", "C_ext"),
  tp_b_mult = NULL
)
}
\arguments{
\item{x}{.}

\item{vary}{.}

\item{tp_b_mult}{.}
}
\value{
.
}
\description{
Title
}
