% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sensitivity.R
\name{compute_sensitivity}
\alias{compute_sensitivity}
\title{Compute response sensitivity to parameter variation.}
\usage{
compute_sensitivity(
  x,
  vary = c("age", "obesity", "css_params", "fit_params", "C_ext"),
  tp_b_mult = NULL
)
}
\arguments{
\item{x}{GeoTox object.}

\item{vary}{which parameter to vary.}

\item{tp_b_mult}{input for \link{calc_concentration_response} step.}
}
\value{
output from \link{calc_concentration_response}
}
\description{
Compute response sensitivity to parameter variation.
}
