% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hill.R
\name{fit_hill}
\alias{fit_hill}
\title{Fit 2- or 3-parameter Hill model}
\usage{
fit_hill(log10_conc, resp, fixed_slope = TRUE)
}
\arguments{
\item{log10_conc}{base-10 log scale concentration}

\item{resp}{response}

\item{fixed_slope}{if TRUE, slope is fixed at 1}
}
\value{
fit and other stats
}
\description{
Fit 2- or 3-parameter Hill model
}
