% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_tox_data}
\alias{geo_tox_data}
\title{GeoTox data}
\format{
A list with items:
\describe{
\item{exposure}{2019 AirToxScreen exposure concentrations for a subset of
chemicals in North Carolina counties}
\item{dose_response}{A subset of data from InvitroDB for the
"LTEA_HepaRG_CYP1A1_up" assay}
\item{age}{County population estimates for 7/1/2019 in North Carolina}
\item{obesity}{CDC PLACES obesity data for North Carolina counties in
2020}
\item{simulated_css}{Simulated steady-state plasma concentrations for
various age groups and obesity status combinations}
\item{boundaries}{County and state boundaries for North Carolina in 2019}
}
}
\usage{
geo_tox_data
}
\description{
Sample data for use in vignettes and function examples. See the
"package_data" vignette for details on how this data was gathered.
}
\keyword{datasets}
