% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTox-package.R
\docType{package}
\name{GeoTox-package}
\alias{GeoTox-package}
\title{GeoTox: Spatiotemporal Mixture Risk Assessment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Connecting spatiotemporal exposure to individual and population-level risk via source-to-outcome continuum modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://niehs.github.io/GeoTox/}
  \item \url{https://github.com/NIEHS/GeoTox}
}

}
\author{
\strong{Maintainer}: Kyle Messier \email{kyle.messier@nih.gov} (\href{https://orcid.org/0000-0001-9508-9623}{ORCID}) [contributor]

Authors:
\itemize{
  \item Skylar Marvel \email{skylar.marvel@nih.gov} (\href{https://orcid.org/0000-0002-2971-9743}{ORCID}) [contributor]
  \item David Reif \email{david.reif@nih.gov} (\href{https://orcid.org/0000-0001-7815-6767}{ORCID}) [contributor]
}

}
\keyword{internal}
