% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoTox.R
\name{GeoTox}
\alias{GeoTox}
\alias{plot.GeoTox}
\title{GeoTox S3 object}
\usage{
GeoTox()

\method{plot}{GeoTox}(x, type = "resp", ...)
}
\arguments{
\item{x}{a GeoTox object}

\item{type}{type of plot}

\item{...}{additional arguments passed to plotting functions}
}
\description{
An S3 object that can be used to help organize the data and results
of a GeoTox analysis. There are several methods that can be used
on a GeoTox object: \link{simulate}, \link{calc} and the plot generic S3 method
detailed below.
}
\seealso{
\link{simulate}, \link{calc}
}
