% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_ECx.R
\name{obj_ECx}
\alias{obj_ECx}
\title{A Generalized concentration addition style solution}
\usage{
obj_ECx(conc_mix, resp, conc, max, AC50)
}
\arguments{
\item{conc_mix}{effective concentration of the mixture in regular space}

\item{resp}{individual chemical responses}

\item{conc}{individual chemical concentrations in regular space}

\item{max}{maximal (asymtotic) response}

\item{AC50}{concentrations of half-maximal response}
}
\value{
objective value
}
\description{
Find the effective concentration of a mixture via an objective function
given the concentrations and inverse, based on a regular space AC50 and
concentrations.
}
