% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_GCA.R
\name{obj_GCA}
\alias{obj_GCA}
\title{Generalized concentration addition objective function}
\usage{
obj_GCA(ln_resp, conc, max, AC50)
}
\arguments{
\item{ln_resp}{natural log of individual chemical responses}

\item{conc}{individual chemical concentrations in regular space}

\item{max}{maximal (asymtotic) responses}

\item{AC50}{concentrations of half-maximal response}
}
\value{
objective value
}
\description{
Use to find the optimal efficacy value based on a regular space AC50 and
concentrations.
}
