% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exposure.R
\name{plot_exposure}
\alias{plot_exposure}
\title{Title}
\usage{
plot_exposure(exposure, region_boundary, group_boundary = NULL, ncol = 2)
}
\arguments{
\item{exposure}{.}

\item{region_boundary}{.}

\item{group_boundary}{.}

\item{ncol}{.}
}
\value{
.
}
\description{
Title
}
