% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resp.R
\name{plot_resp}
\alias{plot_resp}
\title{Plot response data}
\usage{
plot_resp(
  df,
  region_boundary,
  group_boundary = NULL,
  assay_quantiles = c(Median = 0.5),
  summary_quantiles = c(`10th percentile` = 0.1)
)
}
\arguments{
\item{df}{output from \link{resp_quantiles}.}

\item{region_boundary}{"sf" data.frame mapping features to a "geometry"
column. Used to color map regions.}

\item{group_boundary}{"sf" data.frame containing a "geometry" column. Used
to draw outlines around groups of regions.}

\item{assay_quantiles}{named numeric vector of assay quantile labels.}

\item{summary_quantiles}{named numeric vector of summary quantile labels.}
}
\value{
ggplot2 object.
}
\description{
Plot response data
}
