% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Title}
\usage{
plot_sensitivity(x, param = "GCA.Eff", y = "", xlab = param, ylab = "")
}
\arguments{
\item{x}{.}

\item{param}{.}

\item{y}{.}

\item{xlab}{.}

\item{ylab}{.}
}
\value{
.
}
\description{
Title
}
