% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sensitivity_analysis}
\alias{sensitivity_analysis}
\title{Perform sensitivity analysis}
\usage{
sensitivity_analysis(x)
}
\arguments{
\item{x}{GeoTox object}
}
\value{
The same object with additional fields added or updated
}
\description{
Perform sensitivity analysis
}
