% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_boundaries.R
\name{set_boundaries}
\alias{set_boundaries}
\title{Set GeoTox boundaries}
\usage{
set_boundaries(x, region = NULL, group = NULL)
}
\arguments{
\item{x}{GeoTox object.}

\item{region}{"sf" data.frame mapping features to a "geometry" column. Used
when coloring map regions.}

\item{group}{"sf" data.frame containing a "geometry" column. Used to draw
outlines around groups of regions.}
}
\value{
same GeoTox object with boundaries set.
}
\description{
Set GeoTox boundaries
}
\examples{
geoTox <- GeoTox() |> 
  set_boundaries(region = geo_tox_data$boundaries$county,
                 group  = geo_tox_data$boundaries$state)
}
