% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_age.R
\name{simulate_age}
\alias{simulate_age}
\title{Simulate age data}
\usage{
simulate_age(x, n = 1000)
}
\arguments{
\item{x}{data frame containing population data for age groups.}

\item{n}{simulated sample size.}
}
\value{
array of simulated ages.
}
\description{
Simulate age data
}
\examples{
x <- data.frame(AGEGRP = 0:18, TOT_POP = c(sum(1:18), 1:18))
simulate_age(x, 10)
}
