% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_age.R
\name{simulate_age}
\alias{simulate_age}
\title{Simulate ages}
\usage{
simulate_age(x, n = 1000)
}
\arguments{
\item{x}{data frame or list of data frames containing population data for age
groups. Each data frame must contain columns "AGEGRP" and "TOT_POP".}

\item{n}{simulated sample size.}
}
\value{
Array or list of arrays containing simulated ages.
}
\description{
Simulate ages
}
\examples{
# Single data frame
data <- geo_tox_data$age[geo_tox_data$age$FIPS == 37001, ]
ages <- simulate_age(data, 10)

# List of data frames
data <- split(geo_tox_data$age, ~FIPS)
ages <- simulate_age(data, 10)

}
