% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_exposure.R
\name{simulate_exposure}
\alias{simulate_exposure}
\title{Simulate external exposure}
\usage{
simulate_exposure(x, n = 1000)
}
\arguments{
\item{x}{data frame with a "concentration_mean" column.}

\item{n}{simulated sample size.}
}
\value{
matrix of inhalation rates.
}
\description{
Simulate external exposure
}
\examples{
x <- data.frame(concentration_mean = 1:3)
simulate_exposure(x, 10)
}
