% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_exposure.R
\name{simulate_exposure}
\alias{simulate_exposure}
\title{Simulate external exposure}
\usage{
simulate_exposure(
  x,
  expos_mean = "mean",
  expos_sd = "sd",
  expos_label = "casn",
  n = 1000
)
}
\arguments{
\item{x}{data frame or list of data frames.}

\item{expos_mean}{column name of mean values.}

\item{expos_sd}{column name of standard deviations.}

\item{expos_label}{column name of labeling term, required if \code{x} has more
than one row.}

\item{n}{simulated sample size.}
}
\value{
A list of matrices containing inhalation rates. Matrix columns are
named using the values in the \code{expos_label} column for more than one data
frame row.
}
\description{
Simulate external exposure
}
\examples{
data <- split(geo_tox_data$exposure, ~FIPS)

# Single data frame
simulate_exposure(data[[1]], n = 5)

# List of data frames
simulate_exposure(data[1:3], n = 5)

}
