% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inhalation_rate.R
\name{simulate_inhalation_rate}
\alias{simulate_inhalation_rate}
\title{Simulate inhalation rates}
\usage{
simulate_inhalation_rate(ages)
}
\arguments{
\item{ages}{array of ages.}
}
\value{
array of inhalation rates.
}
\description{
Simulate inhalation rates
}
\examples{
simulate_inhalation_rate(c(1, 6, 20))
}
