% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_obesity.R
\name{simulate_obesity}
\alias{simulate_obesity}
\title{Simulate obesity status}
\usage{
simulate_obesity(
  x,
  obes_prev = "OBESITY_CrudePrev",
  obes_sd = "OBESITY_SD",
  obes_label = "FIPS",
  n = 1000
)
}
\arguments{
\item{x}{data frame containing obesity data as a percentage from 0 to 100.}

\item{obes_prev}{column name of prevalence.}

\item{obes_sd}{column name of standard deviation.}

\item{obes_label}{column name of labeling term, required if \code{x} has more than
one row.}

\item{n}{simulated sample size(s).}
}
\value{
List of arrays containing simulated obesity status.
}
\description{
Simulate obesity status
}
\details{
The sample size can be either a single value or a vector the same length as
the number of rows in x. If a single value is provided, the same
sample size is used for all data frames. If a vector is provided, each
element corresponds to the sample size for each row in x.
}
\examples{
# Input has default column names
df <- data.frame(OBESITY_CrudePrev = c(20, 50, 80),
                 OBESITY_SD = c(5, 5, 5),
                 FIPS = letters[1:3])
simulate_obesity(df, n = 5)
# different sample sizes
simulate_obesity(df, n = 5:3)

# Input has custom column names
df <- data.frame(prev = c(20, 50, 80),
                 sd = c(5, 5, 5),
                 label = letters[1:3])
simulate_obesity(df,
                 obes_prev = "prev",
                 obes_sd = "sd",
                 obes_label = "label",
                 n = 5)

}
