% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_obesity.R
\name{simulate_obesity}
\alias{simulate_obesity}
\title{Simulate obesity}
\usage{
simulate_obesity(
  x,
  prev = "OBESITY_CrudePrev",
  sd = "OBESITY_SD",
  label = "FIPS",
  n = 1000
)
}
\arguments{
\item{x}{data frame containing obesity data as a percentage from 0 to 100.}

\item{prev}{column name of prevalence.}

\item{sd}{column name of standard deviation.}

\item{label}{column name of labeling term, required if \code{x} has more than one
row.}

\item{n}{simulated sample size.}
}
\value{
Array or named list of arrays containing simulated obesity status.
}
\description{
Simulate obesity
}
\examples{
simulate_obesity(geo_tox_data$obesity[1, ], n = 4)

simulate_obesity(geo_tox_data$obesity[1:5, ], n = 10)

df <- data.frame(label = letters[1:3], prev = c(20, 50, 80), sd = c(5, 5, 5))
simulate_obesity(df, label = "label", prev = "prev", sd = "sd", n = 10)

}
