% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{Simulate population data}
\usage{
simulate_population(
  x,
  age = NULL,
  obesity = NULL,
  exposure = NULL,
  simulated_css = NULL,
  n = 1000,
  ...
)
}
\arguments{
\item{x}{GeoTox object}

\item{age}{input \code{x} to function \link{simulate_age}. After simulating ages, the
inhalation rate is subsequently calculated using \link{simulate_inhalation_rate}.}

\item{obesity}{input \code{x} to function \link{simulate_obesity}}

\item{exposure}{input \code{x} to function \link{simulate_exposure}}

\item{simulated_css}{input \code{simulated_css} to function \link{sample_Css}}

\item{n}{simulated sample size}

\item{...}{additional arguments passed to other functions}
}
\value{
The same object with simulated fields added
}
\description{
Simulate population data for given input fields
}
\examples{
x <- simulate_population(
  GeoTox(),
  age = split(geo_tox_data$age, ~FIPS)[1:5],
  obesity = geo_tox_data$obesity[1:5, ],
  exposure = split(geo_tox_data$exposure, ~FIPS)[1:5],
  simulated_css = geo_tox_data$simulated_css,
  n = 10
)
}
\seealso{
\link{simulate_age}, \link{simulate_inhalation_rate}, \link{simulate_obesity},
\link{simulate_exposure}, \link{sample_Css}
}
